feat(ui): create topics, restrictions and channel image components
