--- libs/dsb-client-gateway/ui/channels/src/lib/Channel/index.ts
@@ -1,1 +1,0 @@
-export * from './Channel';
