--- libs/dsb-client-gateway/ui/channels/src/lib/ChannelInfo/ChannelInfo.styles.ts
@@ -1,43 +1,0 @@
-import { makeStyles } from 'tss-react/mui';
-import { lighten, alpha, darken } from '@mui/material/styles';
-
-export const useStyles = makeStyles()((theme) => ({
-  root: {
-    background: lighten(theme.palette.background.paper, 0.07),
-    boxShadow: 'none',
-    width: 260,
-    minHeight: 243,
-  },
-  card: {
-    padding: '22px 23px',
-  },
-  cardHeader: {
-    borderBottom: `1px solid ${lighten(theme.palette.background.paper, 0.2)}`,
-    marginBottom: 18,
-    paddingBottom: 14,
-  },
-  cardHeaderText: {
-    fontSize: 14,
-    lineHeight: '18px',
-    color: theme.palette.grey[400],
-    fontFamily: theme.typography.body2.fontFamily,
-  },
-  title: {
-    fontSize: 14,
-    lineHeight: '24px',
-    color: theme.palette.common.white,
-    letterSpacing: '0.4px',
-    fontFamily: theme.typography.body2.fontFamily,
-  },
-  subTitle: {
-    fontSize: 12,
-    lineHeight: '14px',
-    color: theme.palette.grey[400],
-    letterSpacing: '0.4px',
-    fontFamily: theme.typography.body2.fontFamily,
-    maxWidth: 190,
-  },
-  row: {
-    marginBottom: 16,
-  },
-}));
