--- libs/dsb-client-gateway/ui/channels/src/lib/ChannelInfo/index.ts
@@ -1,1 +1,0 @@
-export * from './ChannelInfo';
