+++ libs/dsb-client-gateway/ui/channels/src/lib/components/ChannelImage/ChannelImage.styles.ts
@@ -1,0 +1,14 @@
+import { makeStyles } from 'tss-react/mui';
+
+export const useStyles = makeStyles()((theme) => ({
+  imageWrapper: {
+    width: 62,
+    height: 58,
+    borderRadius: 6,
+    background: theme.palette.background.default,
+    marginBottom: 4,
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+}));
