+++ libs/dsb-client-gateway/ui/channels/src/lib/components/ChannelImage/index.ts
@@ -1,0 +1,1 @@
+export * from './ChannelImage';
