+++ libs/dsb-client-gateway/ui/channels/src/lib/components/ChannelList/index.ts
@@ -1,0 +1,1 @@
+export * from './ChannelList';
