+++ libs/dsb-client-gateway/ui/channels/src/lib/components/RestrictionsViewBox/index.ts
@@ -1,0 +1,1 @@
+export * from './RestrictionsViewBox';
