+++ libs/dsb-client-gateway/ui/channels/src/lib/components/index.ts
@@ -1,0 +1,3 @@
+export * from './ChannelImage';
+export * from './ChannelList';
+export * from './RestrictionsViewBox';
