+++ libs/dsb-client-gateway/ui/channels/src/lib/containers/modals/Details/Details.tsx
@@ -7,33 +7,14 @@ import {
   Box,
   IconButton,
 } from '@mui/material';
-import {
-  CloseButton,
-  Dialog,
-  CopyToClipboard,
-} from '@dsb-client-gateway/ui/core';
+import { CloseButton, Dialog } from '@dsb-client-gateway/ui/core';
 import { didFormatMinifier } from '@dsb-client-gateway/ui/utils';
-import { UpdateChannelDtoType } from '@dsb-client-gateway/dsb-client-gateway-api-client';
-import { ChannelType } from '../../../models/type';
+import { ChannelConnectionType } from '../../../models/channel-connection-type.enum';
+import { getChannelType } from '../../../utils';
+import { ChannelImage, RestrictionsViewBox } from '../../../components';
+import { Topics } from './Topics/Topics';
 import { useDetailsEffects } from './Details.effects';
 import { useStyles } from './Details.styles';
-
-interface ChannelImageProps {
-  type: string;
-}
-
-const ChannelImage: FC<ChannelImageProps> = ({ type }) => {
-  const { classes } = useStyles();
-  const icon =
-    type === UpdateChannelDtoType.pub || type === UpdateChannelDtoType.sub
-      ? '/icons/messaging-light.svg'
-      : '/icons/file-transfer-light.svg';
-  return (
-    <Box className={classes.imageWrapper}>
-      <img width={40} src={icon} alt="channel icon" />
-    </Box>
-  );
-};
 
 export const Details: FC = () => {
   const { classes } = useStyles();
@@ -52,10 +33,7 @@ export const Details: FC = () => {
             <Box className={classes.channelWrapper}>
               <ChannelImage type={channel.type} />
               <Typography className={classes.type}>
-                {channel.type === UpdateChannelDtoType.pub ||
-                channel.type === UpdateChannelDtoType.sub
-                  ? 'Messaging'
-                  : 'File transfer'}
+                {getChannelType(channel.type)}
               </Typography>
               <Typography className={classes.namespace}>
                 {channel.fqcn}
@@ -64,6 +42,7 @@ export const Details: FC = () => {
             <Box className={classes.details}>
               <Typography className={classes.title}>Details</Typography>
               <IconButton
+                disableRipple
                 className={classes.editIconButton}
                 onClick={openUpdateChannel}
               >
@@ -75,7 +54,7 @@ export const Details: FC = () => {
                 Type:
               </Typography>
               <Typography className={classes.typeValue} variant="body2">
-                {ChannelType[channel.type]}
+                {ChannelConnectionType[channel.type]}
               </Typography>
             </Box>
             <Box mb={0.9}>
@@ -84,101 +63,21 @@ export const Details: FC = () => {
               </Typography>
             </Box>
             <Box display="flex">
-              <Box className={classes.restrictionsBox} mr={0.8}>
-                <Box mb={1}>
-                  <Typography
-                    variant="body2"
-                    className={classes.restrictionsBoxLabel}
-                  >
-                    Role
-                  </Typography>
-                </Box>
-                <Box className={classes.restictionsList}>
-                  {channel.conditions?.roles.map((role) => {
-                    return (
-                      <Box key={role} mb={0.4} display="flex">
-                        <Typography
-                          noWrap
-                          variant="body2"
-                          className={classes.restrictionsItemtext}
-                        >
-                          {role}
-                        </Typography>
-                        <CopyToClipboard
-                          text={role}
-                          wrapperProps={{ display: 'flex' }}
-                        />
-                      </Box>
-                    );
-                  })}
-                </Box>
-              </Box>
-              <Box className={classes.restrictionsBox} ml={0.8}>
-                <Box mb={1}>
-                  <Typography
-                    variant="body2"
-                    className={classes.restrictionsBoxLabel}
-                  >
-                    DID
-                  </Typography>
-                </Box>
-                <Box className={classes.restictionsList}>
-                  {channel.conditions?.dids.map((did) => {
-                    return (
-                      <Box key={did} mb={0.4} display="flex">
-                        <Typography
-                          noWrap
-                          variant="body2"
-                          className={classes.restrictionsItemtext}
-                        >
-                          {didFormatMinifier(did, 5, 3)}
-                        </Typography>
-                        <CopyToClipboard
-                          text={did}
-                          wrapperProps={{ display: 'flex' }}
-                        />
-                      </Box>
-                    );
-                  })}
-                </Box>
-              </Box>
+              <RestrictionsViewBox
+                label="Role"
+                list={channel.conditions?.roles}
+                wrapperProps={{ mr: 0.8 }}
+              />
+              <RestrictionsViewBox
+                label="DID"
+                list={channel.conditions?.dids}
+                formatter={(value: string) => didFormatMinifier(value, 5, 3)}
+                wrapperProps={{ ml: 0.8 }}
+              />
             </Box>
             <Box className={classes.divider}></Box>
             <Box mt={2.6}>
-              <Typography className={classes.label} variant="body2">
-                Topics
-              </Typography>
-              <Box className={classes.topicsList}>
-                {channel.conditions?.topics.map((topic) => {
-                  return (
-                    <Box
-                      key={topic.topicId}
-                      display="flex"
-                      flexDirection="column"
-                      mb={0.4}
-                    >
-                      <Typography
-                        className={classes.topicLabel}
-                        variant="body2"
-                      >
-                        {topic.topicName}
-                      </Typography>
-                      <Box display="flex">
-                        <Typography
-                          className={classes.topicValue}
-                          variant="body2"
-                        >
-                          {topic.owner}
-                        </Typography>
-                        <CopyToClipboard
-                          text={topic.owner}
-                          wrapperProps={{ display: 'flex', marginLeft: '10px' }}
-                        />
-                      </Box>
-                    </Box>
-                  );
-                })}
-              </Box>
+              <Topics topics={channel.conditions?.topics} />
             </Box>
           </>
         )}
