+++ libs/dsb-client-gateway/ui/channels/src/lib/containers/modals/Details/Topics/Topics.styles.ts
@@ -1,0 +1,39 @@
+import { makeStyles } from 'tss-react/mui';
+
+export const useStyles = makeStyles()((theme) => ({
+  label: {
+    fontSize: 12,
+    lineHeight: '14px',
+    fontWeight: 400,
+    color: theme.palette.grey[300],
+  },
+  topicLabel: {
+    fontSize: 14,
+    lineHeight: '21px',
+    fontWeight: 400,
+    color: theme.palette.common.white,
+  },
+  topicValue: {
+    fontSize: 14,
+    lineHeight: '21px',
+    fontWeight: 400,
+    color: theme.palette.grey[300],
+  },
+  topicsList: {
+    maxHeight: 90,
+    marginTop: 15,
+    overflow: 'auto',
+    '&::-webkit-scrollbar': {
+      width: 2,
+    },
+    '&::-webkit-scrollbar-track': {
+      backgroundColor: theme.palette.background.paper,
+      boxSizing: 'border-box',
+      borderRadius: 3,
+    },
+    '&::-webkit-scrollbar-thumb': {
+      backgroundColor: theme.palette.primary.main,
+      borderRadius: 3,
+    },
+  },
+}));
