+++ libs/dsb-client-gateway/ui/channels/src/lib/containers/modals/models/create-channel.interface.ts
@@ -1,5 +1,5 @@
 import { CreateChannelDto } from '@dsb-client-gateway/dsb-client-gateway-api-client';
-import { ChannelType } from '../Create/Details/models/channel-type.enum';
+import { ChannelType } from '../../../models/channel-type.enum';
 
 export interface ICreateChannel extends CreateChannelDto {
   connectionType: ConnectionType | string;
