+++ libs/dsb-client-gateway/ui/channels/src/lib/models/channel-connection-type.enum.ts
@@ -1,4 +1,4 @@
-export enum ChannelType {
+export enum ChannelConnectionType {
   pub = 'Publish',
   sub = 'Subscribe',
   download = 'Download',
