+++ libs/dsb-client-gateway/ui/channels/src/lib/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './getChannelTypeImage';
+export * from './getChannelType';
